move config package
