+++ service/src/main/java/fr/funixgaming/api/server/external_api_impl/twitch/auth/services/TwitchClientTokenService.java
@@ -11,12 +11,12 @@ import fr.funixgaming.api.core.exceptions.ApiForbiddenException;
 import fr.funixgaming.api.core.exceptions.ApiNotFoundException;
 import fr.funixgaming.api.core.utils.string.PasswordGenerator;
 import fr.funixgaming.api.server.external_api_impl.twitch.auth.clients.TwitchTokenAuthClient;
-import fr.funixgaming.api.server.external_api_impl.twitch.auth.configs.TwitchApiConfig;
 import fr.funixgaming.api.server.external_api_impl.twitch.auth.dtos.TwitchTokenResponseDTO;
 import fr.funixgaming.api.server.external_api_impl.twitch.auth.dtos.TwitchValidationTokenResponseDTO;
 import fr.funixgaming.api.server.external_api_impl.twitch.auth.entities.TwitchClientToken;
 import fr.funixgaming.api.server.external_api_impl.twitch.auth.mappers.TwitchClientTokenMapper;
 import fr.funixgaming.api.server.external_api_impl.twitch.auth.repositories.TwitchClientTokenRepository;
+import fr.funixgaming.api.server.external_api_impl.twitch.configs.TwitchApiConfig;
 import fr.funixgaming.api.server.user.entities.User;
 import fr.funixgaming.api.server.user.services.UserService;
 import lombok.Getter;
